[expo] add apploading polyfill for bare projects
