+++ packages/expo/src/launch/SplashScreen.ts
@@ -1,15 +1,1 @@
-import { NativeModules } from 'react-native';
-
-const { ExponentSplashScreen: SplashScreen = {} } = NativeModules;
-
-export function preventAutoHide() {
-  if (SplashScreen.preventAutoHide) {
-    SplashScreen.preventAutoHide();
-  }
-}
-
-export function hide() {
-  if (SplashScreen.hide) {
-    SplashScreen.hide();
-  }
-}
+export * from 'expo-splash-screen';
